feat: added empty files where we can re-create the charts of the other repo
