+++ stories/index.js
@@ -1,11 +1,13 @@
 import { storiesOf } from '@storybook/vue'
 
-import Scatterplot from './charts/Scatterplot.vue'
-import BarChart from './charts/BarChart.vue'
-import PlotLines from './charts/PlotLines.vue'
-import TestCategoricalDomain from './charts/TestCategoricalDomain.vue'
+import Scatterplot from './sandbox/Scatterplot.vue'
+import BarChart from './sandbox/BarChart.vue'
+import PlotLines from './sandbox/PlotLines.vue'
+import TestCategoricalDomain from './sandbox/TestCategoricalDomain.vue'
 
-storiesOf('Charts', module)
+// storiesOf('Charts', module)
+
+storiesOf('Sandbox', module)
   .add('Scatterplot', () => (Scatterplot))
   .add('Bar chart', () => (BarChart))
   .add('Plot lines', () => (PlotLines))
