+++ stories/sandbox/TestCategoricalDomain.vue
